Renamed Environment.js to ConfigFile.js
