+++ bin/configFile.js
