+++ bin/tasks/setup_ssl.js
@@ -10,9 +10,9 @@ var path = require('path');
 var inquirer = require('inquirer');
 var askFor = require('./_questionnaire');
 var crypto = require('crypto');
-var environment = require('../environment');
+var configFile = require('../configFile');
 
-var config = _.defaults(environment.getSync(), {
+var config = _.defaults(configFile.getSync(), {
     connections: {
         storage: new Object,
         authentication: new Object
